Separate CLI bin from ota lib
