+++ src/lib.rs
@@ -1,0 +1,23 @@
+extern crate clap;
+extern crate env_logger;
+#[macro_use]
+extern crate log;
+extern crate reqwest;
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+#[macro_use]
+extern crate serde_json;
+extern crate serde_urlencoded;
+extern crate toml;
+extern crate url;
+extern crate url_serde;
+extern crate urlencoding;
+extern crate uuid;
+extern crate zip;
+
+pub mod api;
+pub mod command;
+pub mod config;
+pub mod error;
+pub mod http;
